% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblionetwork-package.R
\docType{package}
\name{biblionetwork-package}
\alias{biblionetwork}
\alias{biblionetwork-package}
\title{biblionetwork: Create Different Types of Bibliometric Networks}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions to find edges for bibliometric networks like bibliographic coupling network, co-citation network and co-authorship network. The weights of network edges can be calculated according to different methods, depending on the type of networks, the type of nodes, and what you want to analyse. These functions are optimized to be be used on large dataset. The package contains functions inspired by: Leydesdorff, Loet and Park, Han Woo (2017) \doi{10.1016/j.joi.2016.11.007}; Perianes-Rodriguez, Antonio, Ludo Waltman, and Nees Jan Van Eck (2016) \doi{10.1016/j.joi.2016.10.006}; Sen, Subir K. and Shymal K. Gan (1983) \url{http://nopr.niscair.res.in/handle/123456789/28008}; Shen, Si, Zhu, Danhao, Rousseau, Ronald, Su, Xinning and Wang, Dongbo (2019) \doi{10.1016/j.joi.2019.01.012}; Zhao, Dangzhi and Strotmann, Andreas (2008) \doi{10.1002/meet.2008.1450450292}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agoutsmedt/biblionetwork}
  \item \url{https://agoutsmedt.github.io/biblionetwork/}
  \item Report bugs at \url{https://github.com/agoutsmedt/biblionetwork/issues}
}

}
\author{
\strong{Maintainer}: Aurélien Goutsmedt \email{agoutsmedt@hotmail.fr} (\href{https://orcid.org/0000-0002-3788-7237}{ORCID})

Authors:
\itemize{
  \item François Claveau \email{françois.claveau@usherbrooke.ca} (\href{https://orcid.org/0000-0001-7129-7794}{ORCID})
  \item Alexandre Truc \email{alexandre.truc77@gmail.com} (\href{https://orcid.org/0000-0002-1328-7819}{ORCID})
}

}
\keyword{internal}
