% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist.R
\name{make_edgelist}
\alias{make_edgelist}
\title{Creates an edgelist}
\usage{
make_edgelist(df, node_cols, separator)
}
\arguments{
\item{df}{dataframe containing the relevant columns}

\item{nodeID}{column containing concatenated string values of co-occuring
node identifiers, e.g., 1 row = nodeID1, nodeID2, nodeID3}

\item{sep}{separator character in \code{nodes_col}, e.g., " , "}
}
\value{
The output is a df with three columns:
X1 = head node, X2 = end node, weight = numeric weight of edge
(these are undirected networks, so edge pairs (i,j) and (j,i) are equivalent).
}
\description{
\code{make_edgelist} generates an edgelist (a co-occurrence list) from
relational data
}
\details{
This generates edgelists for \strong{undirected graphs} from (typically)
bibliometric data It is built off of work by
\href{https://www.r-bloggers.com/turning-keywords-into-a-co-occurrence-network/}{Francois
Briatte}. Input data are usually of two common forms:

1. A co-citation network: The nodes are publications (one node = one
publication), and the edges connect publications that share references, i.e.,
if publication 1 and publication 2 both cite reference A, they are connected
by an edge.

2. A co-author network: The nodes are authors (one node = one author), and
the edges connect authors who wrote a publication together.
}
