% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodelist.R
\name{make_nodelist}
\alias{make_nodelist}
\title{Creates a (distinct) nodelist}
\usage{
make_nodelist(df, nodes_col, separator, keep_col = FALSE, ...)
}
\arguments{
\item{df}{dataframe containing the relevant columns}

\item{nodes_col}{column containing concatenated string values of co-occurring
node identifiers, e.g., 1 row = nodeID1, nodeID2, nodeID3}

\item{separator}{= separator character in \code{nodes_col}, e.g., " , "}

\item{keep_col}{logical, becomes the logical argument for distinct(.keep_all
= )}

\item{...}{other columns to use in distinct() to retain distinct combinations
rather than simply distinct values of \code{nodes_col}}
}
\value{
The default output is one column with distinct nodes only, keeping no
  additional attributes/columns.

  If keep_col = TRUE, keeps all additional attributes in the dataframe.

  If extra arguments (...) are provided, these are additional arguments to
  distinct(); i.e., distinct(nodes_col, ..., .keep_all = keep_col).
}
\description{
\code{make_nodelist} generates a nodelist from relational data for network
graphs
}
\details{
This function, by default, gives a list of \strong{distinct} nodes only.
There is an option for preserving node attributions (i.e., keeping additional
columns in the dataframe), or dropping them.
}
