% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibli_coupling.R
\name{biblio_coupling}
\alias{biblio_coupling}
\title{function for edges of bibliographic coupling}
\usage{
biblio_coupling(
  dt,
  source,
  ref,
  normalized_weight_only = TRUE,
  weight_threshold = 1,
  output_in_character = TRUE
)
}
\arguments{
\item{dt}{For bibliographic coupling, the dataframe with citing and cited documents. It could also be used
\enumerate{
\item for title co-occurence network, with \code{source} being the articles,
and \code{ref} being the list of words in articles titles;
\item for co-authorship network,
with \code{source} being the authors, and \code{ref} the list of articles. For co-authorship,
rather use the \code{coauth_network} function.
}}

\item{source}{the column name of the source identifiers, that is the documents that are citing.}

\item{ref}{the column name of the cited references identifiers.}

\item{normalized_weight_only}{if set to FALSE, the function returns the weights normalized by the cosine measure,
but also simply the number of shared references.}

\item{weight_threshold}{Correspond to the value of the non-normalized weights of edges. The function just keeps the edges
that have a non-normalized weight superior to the \code{weight_threshold}. In a large bibliographic coupling network,
you can consider for instance that sharing only one reference is not sufficient/significant for two articles to be linked together.
This parameter could also be modified to avoid creating untractable networks with too many edges.}

\item{output_in_character}{If TRUE, the function ends by transforming the \code{from} and \code{to} columns in character, to make the
creation of a tidygraph object easier.}
}
\value{
A data.table with the articles (or authors) identifier in \code{from} and \code{to} columns, with one or two additional columns (the coupling angle measure and
the number of shared references). It also keeps a copy of \code{from} and \code{to} in the \code{Source} and \code{Target} columns. This is useful is you
are using the tidygraph package then, where \code{from} and \code{to} values are modified when creating a graph.
}
\description{
This function calculates the number of references that different articles share together, as weel as
the coupling angle value of edges in a bibliographic coupling network (Sen and Gan, 1983), from a direct
citation data frame. This is a standard way to build bibliographic coupling network using Salton's cosine measure:
it divides the number of references that two articles share by the rootsquare of the product of
both articles bibliography lengths. It avoids giving too much importance to articles with a large bibliography.
}
\details{
This function uses data.table package and is thus very fast. It allows the user to compute the coupling angle
on a very large dataframe very quickly.

This function is a relatively general function that can also be used
\enumerate{
\item for title co-occurence networks (taking care of
the lenght of the title thanks to the coupling angle measure);
\item for co-authorship networks (taking care of the
number of co-authors an author has collaborated with on a period)
}
}
\references{
Sen, S. K., & Gan, S. K. (1983). A mathematical extension of the idea of bibliographic coupling and its applications.
}
