% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupling_entity.R
\name{coupling_entity}
\alias{coupling_entity}
\title{function for edges of bibliographic coupling}
\usage{
coupling_entity(
  dt,
  source,
  ref,
  entity,
  weight_threshold = 1,
  output_in_character = FALSE,
  method = c("coupling_strength", "coupling_angle")
)
}
\arguments{
\item{dt}{The table with citing and cited documents.}

\item{source}{the column name of the source identifiers, that is the documents that are citing.}

\item{ref}{the column name of the cited references identifiers.}

\item{entity}{the column name of the entity (authors, journals, institutions) that are citing.}

\item{weight_threshold}{Correspond to the value of the non-normalized weights of edges. The function just keeps the edges
that have a non-normalized weight superior to the \code{weight_threshold}. In a large bibliographic coupling network,
you can consider for instance that sharing only one reference is not sufficient/significant for two articles (above all for two entities like authors or journals)
to be linked together. This parameter could also be modified to avoid creating untractable networks with too many edges.}

\item{output_in_character}{If TRUE, the function ends by transforming the \code{from} and \code{to} columns in character, to make the
creation of a \href{https://tidygraph.data-imaginist.com/index.html}{tidygraph} graph easier.}

\item{method}{Choose the method you want to use for calculating the edges weights: either "coupling_strength" like in the \code{\link[=coupling_strength]{coupling_strength()}} function,
or "coupling_angle" like in the \code{\link[=biblio_coupling]{biblio_coupling()}} function.}
}
\value{
A data.table with the entity identifiers in \code{from} and \code{to} columns, with the coupling strength or coupling angle measures in
another column, as well as the method used. It also keeps a copy of \code{from} and \code{to} in the \code{Source} and \code{Target} columns. This is useful is you
are using the tidygraph package then, where \code{from} and \code{to} values are modified when creating a graph.
}
\description{
This function creates the edges of a network of entities from a direct citations data frame (i.e. documents citing references).
Entities could be authors, affiliations, journals, etc. Consequently, Coupling links are calculated using the coupling angle measure or the coupling strength measure.
Coupling links are calculated depending of the number references two authors share, taking into account the minimum number of times two authors are citing
each references. For instance, if two entities share a reference in common, the first one citing it twice (in other words, citing it in two different articles),
the second one three times, the function takes two as the minimum value. In addition to the features of the coupling strength measure (see \code{\link[=coupling_strength]{coupling_strength()}})
or the coupling angle measure (see \code{\link[=biblio_coupling]{biblio_coupling()}}), it means that, if two entities share two reference in common, the fact that the first reference is cited
at least four times by the two entities, whereas the second reference is cited at least only once, the first reference contributes more to the edge weight than
the second reference. This use of minimum shared reference for entities coupling comes from \insertCite{zhao2008b;textual}{biblionetwork}.
}
\references{
\insertAllCited{}
}
