% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupling_strength.R
\name{coupling_strength}
\alias{coupling_strength}
\title{Calculating The Coupling Strength Measure For Edges}
\usage{
coupling_strength(
  dt,
  source,
  ref,
  weight_threshold = 1,
  output_in_character = FALSE
)
}
\arguments{
\item{dt}{The table with citing and cited documents.}

\item{source}{the column name of the source identifiers, that is the documents that are citing.}

\item{ref}{the column name of the references that are cited.}

\item{weight_threshold}{Correspond to the value of the non-normalized weights of edges. The function just keeps the edges
that have a non-normalized weight superior to the \code{weight_threshold}. In a large bibliographic coupling network,
you can consider for instance that sharing only one reference is not sufficient/significant for two articles to be linked together.
This parameter could also be modified to avoid creating untractable networks with too many edges.}

\item{output_in_character}{If TRUE, the function ends by transforming the \code{from} and \code{to} columns in character, to make the
creation of a \href{https://tidygraph.data-imaginist.com/index.html}{tidygraph} graph easier.}
}
\value{
A data.table with the articles identifiers in \code{from} and \code{to} columns, with the coupling strength measure in
another column. It also keeps a copy of \code{from} and \code{to} in the \code{Source} and \code{Target} columns. This is useful is you
are using the tidygraph package then, where \code{from} and \code{to} values are modified when creating a graph.
}
\description{
This function calculates the coupling strength measure \insertCite{@following @vladutz1984 and @shen2019}{biblionetwork}
from a direct citation data frame. It is a refinement of \link[=biblio_coupling]{biblio_coupling()}:
it takes into account the frequency with which a reference shared by two articles has been cited in the whole corpus.
In other words, the most cited references are less important in the links between two articles, than references that have
been rarely cited. To a certain extent, it is similar to the TF-IDF measure.
}
\references{
\insertAllCited{}
}
